Moving banner to top of menu
