--- src/styles/_txLimitBanner.scss
@@ -1,19 +1,0 @@
-#tx-limit-banner {
-  background: linear-gradient(45deg, #C2449F, #422453);
-  z-index: 1020;
-  width: 100%;
-
-  div:nth-child(2) {
-    font-size: 16px;
-  }
-  div:last-child {
-    img {
-      width: 105px;
-      height: auto;
-      position: absolute;
-      left: -17px;
-      opacity: 0.1;
-      bottom: -60px;
-    }
-  }
-}
