+++ src/styles/application.scss
@@ -24,7 +24,7 @@
 @import 'createMilestone.scss';
 // components
 @import 'loader.scss';
-@import 'mainMenu.scss';
+@import 'header.scss';
 @import 'goBackButton.scss';
 @import 'loaderButton.scss';
 @import 'joinGivethCommunity.scss';
@@ -32,7 +32,6 @@
 @import 'dac-card.scss';
 @import 'milestoneConversations.scss';
 @import 'conversationModal.scss';
-@import 'txLimitBanner.scss';
 // misc
 @import 'quill-overrides.scss';
 @import 'toastify-overrides.scss';
